# cmd/py_faas/api.gen.go
// Package main provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.5.0 DO NOT EDIT.
package main

import (
	"fmt"
	"net/http"

	"github.com/go-chi/chi/v5"
	"github.com/oapi-codegen/runtime"
	openapi_types "github.com/oapi-codegen/runtime/types"
)

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	// Code Error code
	Code string `json:"code"`

	// Error Error message
	Error string `json:"error"`
}

// GetUserMcpServicesParams defines parameters for GetUserMcpServices.
type GetUserMcpServicesParams struct {
	// User User identifier
	User string `form:"user" json:"user"`
}

// ProcessFileMultipartBody defines parameters for ProcessFile.
type ProcessFileMultipartBody struct {
	// File The file to be processed
	File openapi_types.File `json:"file"`

	// User User identifier
	User string `json:"user"`
}

// ProxyPostRequestJSONBody defines parameters for ProxyPostRequest.
type ProxyPostRequestJSONBody = map[string]interface{}

// ProxyPostRequestMultipartBody defines parameters for ProxyPostRequest.
type ProxyPostRequestMultipartBody = map[string]interface{}

// ProxyPostRequestTextBody defines parameters for ProxyPostRequest.
type ProxyPostRequestTextBody = string

// ProcessFileMultipartRequestBody defines body for ProcessFile for multipart/form-data ContentType.
type ProcessFileMultipartRequestBody ProcessFileMultipartBody

// ProxyPostRequestJSONRequestBody defines body for ProxyPostRequest for application/json ContentType.
type ProxyPostRequestJSONRequestBody = ProxyPostRequestJSONBody

// ProxyPostRequestMultipartRequestBody defines body for ProxyPostRequest for multipart/form-data ContentType.
type ProxyPostRequestMultipartRequestBody = ProxyPostRequestMultipartBody

// ProxyPostRequestTextRequestBody defines body for ProxyPostRequest for text/plain ContentType.
type ProxyPostRequestTextRequestBody = ProxyPostRequestTextBody

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get MCP services for a user
	// (GET /mcp-service)
	GetUserMcpServices(w http.ResponseWriter, r *http.Request, params GetUserMcpServicesParams)
	// Process file input
	// (POST /mcp-service)
	ProcessFile(w http.ResponseWriter, r *http.Request)
	// Proxy GET request to underlying service
	// (GET /{service-uuid}/mcp)
	ProxyGetRequest(w http.ResponseWriter, r *http.Request, serviceUUID openapi_types.UUID)
	// Proxy POST request to underlying service
	// (POST /{service-uuid}/mcp)
	ProxyPostRequest(w http.ResponseWriter, r *http.Request, serviceUUID openapi_types.UUID)
}

// Unimplemented server implementation that returns http.StatusNotImplemented for each endpoint.

type Unimplemented struct{}

// Get MCP services for a user
// (GET /mcp-service)
func (_ Unimplemented) GetUserMcpServices(w http.ResponseWriter, r *http.Request, params GetUserMcpServicesParams) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Process file input
// (POST /mcp-service)
func (_ Unimplemented) ProcessFile(w http.ResponseWriter, r *http.Request) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Proxy GET request to underlying service
// (GET /{service-uuid}/mcp)
func (_ Unimplemented) ProxyGetRequest(w http.ResponseWriter, r *http.Request, serviceUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// Proxy POST request to underlying service
// (POST /{service-uuid}/mcp)
func (_ Unimplemented) ProxyPostRequest(w http.ResponseWriter, r *http.Request, serviceUUID openapi_types.UUID) {
	w.WriteHeader(http.StatusNotImplemented)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetUserMcpServices operation middleware
func (siw *ServerInterfaceWrapper) GetUserMcpServices(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetUserMcpServicesParams

	// ------------- Required query parameter "user" -------------

	if paramValue := r.URL.Query().Get("user"); paramValue != "" {

	} else {
		siw.ErrorHandlerFunc(w, r, &RequiredParamError{ParamName: "user"})
		return
	}

	err = runtime.BindQueryParameter("form", true, true, "user", r.URL.Query(), &params.User)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "user", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetUserMcpServices(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ProcessFile operation middleware
func (siw *ServerInterfaceWrapper) ProcessFile(w http.ResponseWriter, r *http.Request) {

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ProcessFile(w, r)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ProxyGetRequest operation middleware
func (siw *ServerInterfaceWrapper) ProxyGetRequest(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "service-uuid" -------------
	var serviceUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "service-uuid", chi.URLParam(r, "service-uuid"), &serviceUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "service-uuid", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ProxyGetRequest(w, r, serviceUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// ProxyPostRequest operation middleware
func (siw *ServerInterfaceWrapper) ProxyPostRequest(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "service-uuid" -------------
	var serviceUUID openapi_types.UUID

	err = runtime.BindStyledParameterWithOptions("simple", "service-uuid", chi.URLParam(r, "service-uuid"), &serviceUUID, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "service-uuid", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.ProxyPostRequest(w, r, serviceUUID)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{})
}

type ChiServerOptions struct {
	BaseURL          string
	BaseRouter       chi.Router
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, r chi.Router) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseRouter: r,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, r chi.Router, baseURL string) http.Handler {
	return HandlerWithOptions(si, ChiServerOptions{
		BaseURL:    baseURL,
		BaseRouter: r,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options ChiServerOptions) http.Handler {
	r := options.BaseRouter

	if r == nil {
		r = chi.NewRouter()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}
	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/mcp-service", wrapper.GetUserMcpServices)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/mcp-service", wrapper.ProcessFile)
	})
	r.Group(func(r chi.Router) {
		r.Get(options.BaseURL+"/{service-uuid}/mcp", wrapper.ProxyGetRequest)
	})
	r.Group(func(r chi.Router) {
		r.Post(options.BaseURL+"/{service-uuid}/mcp", wrapper.ProxyPostRequest)
	})

	return r
}